+++ apps/frontend/src/app/services/api/search-service.spec.ts
