+++ apps/frontend/src/app/services/api/search-service.ts
