+++ apps/frontend/src/app/services/api/trpc-service.ts
@@ -8,7 +8,7 @@ import { observable } from '@trpc/server/observable';
 import { get, set } from 'idb-keyval';
 
 import { TRPCRouter } from '../../../../backend/src/app/trpc-routers';
-import { environment } from '../../environments/environment';
+import { environment } from '../../../environments/environment';
 import { TokenService } from './token-service';
 import { refreshLink } from './trpc-refreshlink';
