+++ apps/frontend/src/app/services/error.service.ts
@@ -4,7 +4,7 @@ import { JSendServerError } from '@common';
 import { TRPCClientError } from '@trpc/client';
 import { AlertService } from '@uxcommon/alerts/alert-service';
 
-import { TokenService } from '../backend-svc/token-service';
+import { TokenService } from '@services/api/token-service';
 
 /**
  * Centralised error handling service.
