+++ apps/frontend/src/app/uxcommon/datagrid/datagrid.datasource.ts
@@ -3,7 +3,7 @@ import type { getAllOptionsType } from '@common';
 import type { GridApi, IServerSideDatasource, IServerSideGetRowsParams } from 'ag-grid-community';
 
 import type { AbstractAPIService } from '../../abstract-api.service';
-import type { SearchService } from 'apps/frontend/src/app/backend-svc/search-service';
+import type { SearchService } from '@services/api/search-service';
 
 export function createServerSideDatasource<T>(deps: {
   api: GridApi;
